refactor(graphql-shapes): replace patterns query in GtfsMap w/ encoded polys
