+++ lib/common/util/gtfs.js
@@ -1,6 +1,9 @@
 // @flow
 
 import moment from 'moment'
+import {decode as decodePolyline} from 'polyline'
+
+type GraphQLShape = {polyline: string, shape_id: string}
 
 /**
  * @param  {number} seconds Seconds after midnight
@@ -35,4 +38,17 @@ export function secondsAfterMidnightToHHMM (seconds: ?(number | string)): string
  */
 export function humanizeSeconds (seconds: number): string {
   return moment.duration(seconds, 'seconds').humanize()
+}
+
+/**
+ * Array map function to decode a GraphQL encoded shape polyline.
+ */
+export function decodeShapePolylines (shape: GraphQLShape) {
+  return {
+    id: shape.shape_id,
+    // Decode polyline and coords divide by ten (gtfs-lib
+    // simplification level requires this).
+    latLngs: decodePolyline(shape.polyline)
+      .map(coords => ([coords[0] / 10, coords[1] / 10]))
+  }
 }
