+++ lib/editor/actions/tripPattern.js
@@ -9,6 +9,7 @@ import {createVoidPayloadAction, fetchGraphQL, secureFetch} from '../../common/a
 import {snakeCaseKeys} from '../../common/util/map-keys'
 import {generateUID} from '../../common/util/util'
 import {showEditorModal} from './editor'
+import {shapes} from '../../gtfs/util/graphql'
 import {fetchGTFSEntities, receiveGTFSEntities} from '../../manager/actions/versions'
 import {fetchTripCounts} from './trip'
 import {getEditorNamespace} from '../util/gtfs'
@@ -122,27 +123,13 @@ export function fetchTripPatterns (feedId: string) {
   return function (dispatch: dispatchFn, getState: getStateFn) {
     dispatch(fetchingTripPatterns())
     const namespace = getEditorNamespace(feedId, getState())
-    const query = `
-query shapesQuery ($namespace: String) {
-  feed(namespace: $namespace) {
-    feed_version
-    feed_info {
-      feed_id
-    }
-    shapes_as_polylines {
-      shape_id
-      polyline
-    }
-  }
-}
-`
     if (!namespace) {
       console.error('Cannot fetch GTFS for undefined or null namespace')
       dispatch(showEditorModal())
       return
     }
     const variables = {namespace}
-    return dispatch(fetchGraphQL({query, variables}))
+    return dispatch(fetchGraphQL({query: shapes, variables}))
       .then(data => dispatch(
         receiveGTFSEntities({
           namespace,
