+++ lib/editor/reducers/data.js
@@ -2,11 +2,11 @@
 
 import update from 'react-addons-update'
 import clone from 'lodash/cloneDeep'
-import {decode as decodePolyline} from 'polyline'
 import SortDirection from 'react-virtualized/dist/commonjs/Table/SortDirection'
 
-import {ENTITY} from '../constants'
+import {decodeShapePolylines} from '../../common/util/gtfs'
 import {defaultSorter} from '../../common/util/util'
+import {ENTITY} from '../constants'
 import {generateNullProps, getTableById, getKeyForId} from '../util/gtfs'
 import {getMapToGtfsStrategy, entityIsNew} from '../util/objects'
 import {assignDistancesToPatternStops, constructShapePoints} from '../util/map'
@@ -246,15 +246,7 @@ const data = (state: DataState = defaultState, action: Action): DataState => {
         // Handle trip patterns to be drawn as overlay layer in editor
         return update(state, {
           tripPatterns: {$set:
-            data.feed.shapes_as_polylines.map(shape => {
-              return {
-                id: shape.shape_id,
-                // Decode polyline and coords divide by ten (gtfs-lib
-                // simplification level requires this).
-                latLngs: decodePolyline(shape.polyline)
-                  .map(coords => ([coords[0] / 10, coords[1] / 10]))
-              }
-            })
+            data.feed.shapes_as_polylines.map(decodeShapePolylines)
           }
         })
       }
