+++ lib/gtfs/actions/shapes.js
@@ -3,6 +3,7 @@
 import {createAction, type ActionType} from 'redux-actions'
 
 import {createVoidPayloadAction, fetchGraphQL} from '../../common/actions'
+import {decodeShapePolylines} from '../../common/util/gtfs'
 import {shapes} from '../../gtfs/util/graphql'
 import {updateRoutesOnMapDisplay} from './filter'
 
@@ -34,13 +35,8 @@ export function toggleShowAllRoutesOnMap (namespace: string) {
       dispatch(fetchingShapes())
       return dispatch(fetchGraphQL({query: shapes, variables: {namespace}}))
         .then(data => {
-          const {patterns} = data.feed
-          const shapes = patterns.map(pattern => {
-            return pattern.shape.map(point => [
-              point.shape_pt_lat,
-              point.shape_pt_lon
-            ])
-          })
+          const {shapes_as_polylines: encodedShapes} = data.feed
+          const shapes = encodedShapes.map(decodeShapePolylines)
           dispatch(receiveShapes(shapes))
         })
     }
