+++ lib/gtfs/components/GtfsMap.js
@@ -1,8 +1,8 @@
 // @flow
 
-import React, { Component } from 'react'
+import React, { PureComponent } from 'react'
 import { shallowEqual } from 'react-pure-render'
-import { latLngBounds } from 'leaflet'
+import L, { latLngBounds } from 'leaflet'
 import { FeatureGroup, Map, Polyline, Rectangle, TileLayer } from 'react-leaflet'
 
 import * as filterActions from '../actions/filter'
@@ -13,7 +13,7 @@ import PatternGeoJson from './PatternGeoJson'
 import StopMarker from './StopMarker'
 
 import type {Props as ContainerProps} from '../containers/ActiveGtfsMap'
-import type {Bounds, Feed, LatLng, StopWithFeed} from '../../types'
+import type {Bounds, Feed, StopWithFeed} from '../../types'
 import type {DateTimeFilter, MapFilter} from '../../types/reducers'
 
 type Props = ContainerProps & {
@@ -38,26 +38,11 @@ type Props = ContainerProps & {
   updateMapState: typeof filterActions.updateMapState,
 }
 
-type State = {
-  bounds: any,
-  lastClicked?: LatLng,
-  map: any,
-  searchFocus?: any
-}
+const CANVAS = L.canvas()
 
-const DEFAULT_BOUNDS = [[70, 130], [-70, -130]]
-
-export default class GtfsMap extends Component<Props, State> {
+export default class GtfsMap extends PureComponent<Props> {
   static defaultProps = {
     entities: []
-  }
-
-  componentWillMount () {
-    this.setState({
-      bounds: this.props.bounds || DEFAULT_BOUNDS,
-      map: {},
-      searchFocus: null
-    })
   }
 
   componentDidMount () {
@@ -90,7 +75,6 @@ export default class GtfsMap extends Component<Props, State> {
       entities,
       feeds,
       height,
-      searchFocus,
       stop,
       width
     } = this.props
@@ -99,9 +83,6 @@ export default class GtfsMap extends Component<Props, State> {
         // Refresh elements if feeds or entities to search change.
         this._refreshGtfsElements(nextProps.feeds)
       }
-    }
-    if (nextProps.searchFocus && nextProps.searchFocus !== searchFocus) {
-      this.setState({searchFocus: nextProps.searchFocus})
     }
     if (nextProps.stop && !shallowEqual(nextProps.stop, stop)) {
       // handle stop: panning on stop select
@@ -169,8 +150,7 @@ export default class GtfsMap extends Component<Props, State> {
   }
 
   layerAddHandler = (e: any) => {
-    const {stop, pattern} = this.props
-    const {searchFocus} = this.state
+    const {searchFocus, stop, pattern} = this.props
     // handle pattern panTo and popup open
     if (pattern && pattern.pattern_id === searchFocus) {
       // Wait to call open pop up to ensure that layer has fully materialized
@@ -182,12 +162,10 @@ export default class GtfsMap extends Component<Props, State> {
         // Open popup for pattern if added via a selection from the search
         e.layer.openPopup && e.layer.openPopup()
       }, 100)
-      this.setState({searchFocus: null})
     }
     if (stop && searchFocus === stop.stop_id) {
       // Open popup for stop if added via a selection from the search
       e.layer.openPopup && e.layer.openPopup()
-      this.setState({searchFocus: null})
     }
   }
 
@@ -250,6 +228,7 @@ export default class GtfsMap extends Component<Props, State> {
         {showBounds && versionBoundingBox &&
           <Rectangle
             bounds={versionBoundingBox}
+            interactive={false}
             fillOpacity={0} />
         }
 
@@ -330,9 +309,11 @@ export default class GtfsMap extends Component<Props, State> {
           shapes.data.map((shape, index) =>
             <Polyline
               color='#556c7f'
-              key={index}
-              weight={2}
-              positions={shape} />
+              interactive={false}
+              key={shape.id}
+              positions={shape.latLngs}
+              renderer={CANVAS}
+              weight={2} />
           )}
       </Map>
     )
