+++ lib/gtfs/components/ShowAllRoutesOnMapFilter.js
@@ -30,23 +30,27 @@ export default class ShowAllRoutesOnMapFilter extends Component<Props> {
       showAllRoutesOnMap,
       version
     } = this.props
-    // Disable if greater than one million records. We don't want to break
-    // Data Tools because of this shapes GraphQL fetch. This is a somewhat
-    // arbitrary limit. It may need to be adjusted...
+    // If there are over a million records, let the user know that UI
+    // performance might suffer (server should be fine). This is a somewhat
+    // arbitrary limit. NL has over 4M shape records.
     const tooManyShapeRecords = version.feedLoadResult.shapes.rowCount > 1000000
     return (
       <Checkbox
         checked={showAllRoutesOnMap}
-        disabled={tooManyShapeRecords || fetchStatus.fetching}
+        disabled={fetchStatus.fetching}
         onChange={this.onChange}
       >
         {fetchStatus.fetching
           ? <span className='text-muted'>
             {this.messages('fetching')} <Icon className='fa-spin' type='refresh' />
           </span>
-          : tooManyShapeRecords
-            ? this.messages('tooManyShapeRecords')
-            : this.messages('showAllRoutesOnMap')}
+          : <span>
+            {this.messages('showAllRoutesOnMap')}
+            {tooManyShapeRecords && !fetchStatus.fetched &&
+              <small>{' '}({this.messages('tooManyShapeRecords')})</small>
+            }
+          </span>
+        }
       </Checkbox>
     )
   }
