+++ lib/gtfs/util/graphql.js
@@ -194,14 +194,11 @@ query timetablesQuery(
 }`
 
 export const shapes = `
-query shapesQuery($namespace: String) {
-  feed (namespace: $namespace) {
-    patterns(limit: -1) {
-      shape(limit: -1) {
-        shape_pt_lat
-        shape_pt_lon
-        shape_pt_sequence
-      }
+query shapesQuery ($namespace: String) {
+  feed(namespace: $namespace) {
+    shapes_as_polylines {
+      shape_id
+      polyline
     }
   }
 }
